refactor(scrcpy): separate control client to two transform streams
refs #427
