--- libraries/scrcpy/src/control/clipboard.ts
@@ -1,6 +1,0 @@
-import Struct from '@yume-chan/struct';
-
-export const ClipboardMessage =
-    new Struct()
-        .uint32('length')
-        .string('content', { lengthField: 'length' });
