+++ libraries/scrcpy/src/control/index.ts
@@ -1,7 +1,6 @@
-export * from './client.js';
-export * from './clipboard.js';
 export * from './inject-keycode.js';
 export * from './inject-text.js';
 export * from './inject-touch.js';
-export * from './simple.js';
-export * from './types.js';
+export * from './rotate-device.js';
+export * from './stream.js';
+export * from './type.js';
