+++ libraries/scrcpy/src/control/inject-keycode.ts
@@ -1,6 +1,6 @@
 import Struct, { placeholder } from '@yume-chan/struct';
 
-import { ScrcpySimpleControlMessage } from './simple.js';
+import { ScrcpyControlMessageType } from './type.js';
 
 export enum AndroidKeyEventAction {
     Down = 0,
@@ -49,7 +49,7 @@ export enum AndroidKeyCode {
 
 export const ScrcpyInjectKeyCodeControlMessage =
     new Struct()
-        .fields(ScrcpySimpleControlMessage)
+        .uint8('type', ScrcpyControlMessageType.InjectKeyCode as const)
         .uint8('action', placeholder<AndroidKeyEventAction>())
         .uint32('keyCode')
         .uint32('repeat')
