+++ libraries/scrcpy/src/control/rotate-device.ts
@@ -1,0 +1,10 @@
+import Struct from '@yume-chan/struct';
+
+import { ScrcpyControlMessageType } from './type.js';
+
+export const ScrcpyRotateDeviceControlMessage =
+    new Struct()
+        .uint8('type', ScrcpyControlMessageType.RotateDevice as const)
+
+export type ScrcpyRotateDeviceControlMessage =
+    typeof ScrcpyRotateDeviceControlMessage['TInit'];
