--- libraries/scrcpy/src/control/simple.ts
@@ -1,5 +1,0 @@
-import Struct from '@yume-chan/struct';
-
-export const ScrcpySimpleControlMessage =
-    new Struct()
-        .uint8('type');
