+++ libraries/scrcpy/src/device-message/index.ts
@@ -1,0 +1,2 @@
+export * from './clipboard.js';
+export * from './stream.js';
